Fix build errors on linux

+++ sphynx/SphynxServer.cpp
@@ -336,7 +336,7 @@ bool Connection::OnTick(u64 nowMsec)
         return true; // Remove from list
     }
 
-    if (IsFullConnection && nowMsec - LastUDPTimeSyncMsec > S2CUDPTimeSyncIntervalMsec)
+    if (IsFullConnection && nowMsec - LastUDPTimeSyncMsec > static_cast<u64>(S2CUDPTimeSyncIntervalMsec))
     {
         LastUDPTimeSyncMsec = nowMsec;
         LOG(DEBUG) << "Sending UDP timesync " << nowMsec;
